Change GameID to use an int instead of a long.  I'm sure an int is large enough for any length of game, and we save a few bytes on serialization anyway.

+++ PolyPlane/WorldParams.cs
@@ -122,7 +122,7 @@ namespace PolyPlane
         public static readonly D2DPoint PlaneSpawnRange = new D2DPoint(-60000, 60000);
         public static readonly D2DPoint FieldXBounds = new D2DPoint(-350000, 350000);
 
-        public static long CurrentObjId = 0;
+        public static int CurrentObjId = 0;
         public static int CurrentPlayerId = 1000;
 
         public static GameID ViewID;
@@ -201,7 +201,7 @@ namespace PolyPlane
                 TimeOfDayDir = 1f;
         }
 
-        public static long GetNextObjectId()
+        public static int GetNextObjectId()
         {
             return Interlocked.Increment(ref CurrentObjId);
         }
